# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MMSv2",
        "description": "Monospace Member System API 文件",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users/categories": {
            "get": {
                "tags": [
                    "UserCategory"
                ],
                "summary": "Display all users categories.",
                "operationId": "9f1e34fea5cfdce8c144d678fcff613a",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Number of users per page, default is 10",
                        "required": false,
                        "example": "10"
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter userCategory by type name",
                        "required": false,
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UserCategory"
                ],
                "summary": "Create new user category.",
                "operationId": "b4ea18082958df206b71081a1dd95c94",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/categories/{id}": {
            "get": {
                "tags": [
                    "UserCategory"
                ],
                "summary": "Display Category",
                "operationId": "90be3f908ffd971f4b63fc9c00e4d189",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UserCategory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "UserCategory"
                ],
                "summary": "Update the User Category name.",
                "operationId": "15d7e5797fa24c9a29d53233ab72e3cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UserCategory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UserCategory"
                ],
                "summary": "Remove User Category.",
                "operationId": "2a33cbf1ff43f5e57a4d24d38fc40b73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UserCategory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "406": {
                        "description": "Only Accept application/json"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}